+++ implementations/rest/bin/tokens.py
@@ -1,0 +1,10 @@
+#define token functions for substitution in endpoint URL
+# /someurl/foo/$sometoken$/goo -> /someurl/foo/zoo/goo
+
+def sometoken():
+    return 'zoo'
+
+def datetoday():
+    today = datetime.date.today()
+    return today.strftime('%Y-%m-%d')
+